requirements.txt

fastapi
uvicorn
pytest
requests
httpx
psycopg2-binary